pyproject.toml
[project]
name = "cc-wp-book"
version = "0.1.0"
description = "Pipeline to transform Wikipedia's 1,000 Vital Articles into print-ready PDF encyclopedia volumes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "mwparserfromhell>=0.6",
    "weasyprint>=60.0",
    "pillow>=10.0",
    "qrcode>=7.4",
    "pypdf>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "responses>=0.25",
    "ruff>=0.4",
]

[project.scripts]
cc-wp-book = "cc_wp_book.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cc_wp_book"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
